+++ code/modules/events/gimmick/kudzu.dm
@@ -1,4 +1,4 @@
-#define KUDZU_TO_SPREAD_INITIAL 50
+#define KUDZU_TO_SPREAD_INITIAL 40
 /obj/item/kudzuseed//TODO: Move all this to respective files everything works right.
 	name = "kudzu seed"
 	desc = "So this is where Kudzu went. Plant on a floor to grow.<br/>The disclaimer seems faded out, though."
@@ -104,6 +104,7 @@
 	anchored = 1
 	density = 0
 	event_handler_flags = USE_FLUID_ENTER | USE_CANPASS
+	var/static/ideal_temp = 310		//same as blob, why not? I have no other reference point.
 	var/growth = 0
 	var/waittime = 40
 	var/run_life = 0 // I think we have some that spawns on the map so don't just default to growy stuff
@@ -112,6 +113,7 @@
 	var/current_stage = 0
 	var/aggressive = 0
 	var/to_spread = 10				//bascially the radius of child kudzu plants that any given kudzu object can create.
+
 
 	get_desc()
 		var/flavor
@@ -186,11 +188,9 @@
 
 		dmg *= isnum(W.force) ? min((W.force / 2), 5) : 1
 		DEBUG_MESSAGE("[user] damaging [src] with [W] [log_loc(src)]: dmg is [dmg]")
-		src.growth -= dmg
-		if (src.growth < 1)
-			qdel (src)
-		else
-			src.update_self()
+
+		src.take_damage(dmg, "brute", user)
+
 		user.lastattacked  = src
 		..()
 
@@ -315,8 +315,24 @@
 		else
 	return
 
-/obj/spacevine/temperature_expose(null, temp, volume)
-	qdel(src)
+/obj/spacevine/proc/take_damage(var/amount, var/damtype = "brute",var/mob/user)
+	if (!isnum(amount) || amount <= 0)
+		return
+
+	src.growth -= amount
+	if (src.growth < 1)
+		qdel (src)
+	else
+		src.update_self()
+
+
+/obj/spacevine/temperature_expose(datum/gas_mixture/air, temperature, volume)
+	var/temp_diff = temperature - src.ideal_temp
+
+	if (temp_diff >= 300)
+		var/power = max(round(temp_diff /300), 5)
+
+		src.take_damage(power*10, 1, "burn")
 
 /obj/spacevine/living // these ones grow
 	run_life = 1
